// src/types/Page.ts
export type Background = Custom | Standard | Swoosh | Container;
// ...
export type ToggleControl = boolean;
// ...
export interface Image {
  kind?: "image";
  title?: string | null;
  src: string;
  width: number | string;
  height: number | string;
  alt?: string;
  hint?: string;
  is_decorative?: boolean;
  typeOptions?: {
    [k: string]: unknown;
  };
  sizes?: Size[];
  image_background_circle_color?: string;
  image_background_gradient_color?: string;
  image_alignment?: string;
  file_url?: string;
}
export interface Size {
  src?: string;
  width?: number;
  height?: number;
  [k: string]: unknown;
}
// ...
export interface Cta {
  /**
   * The label of the button.
   */
  label?: string;
  ariaLabel?: string;
  url?: string;
  /**
   * Icon identifier
   */
  icon?: string;
  variant?: "primary" | "secondary" | "tertiary" | "text" | "trial";
  target?: "_self" | "_parent" | "_top" | "_blank";
  id?: string | number;
  source?: "standard-link" | "modal-link" | "custom-action-link";
  custom_action_options?: {
    event_code?: string;
    workflow_uri?: string;
    name_capture_campaign_id?: string;
    action_template?: "everybody-demo-org" | "event-sign-up" | "offer";
    [k: string]: unknown;
  };
  linkAttributes?: {
    key?: string;
    value?: string;
    [k: string]: unknown;
  }[];
  modal_variant?: string;
  [k: string]: unknown;
}
// ...
export interface Eyebrow {
  label?: string;
  image?: Image;
  icon?: Icon;
  info_bubble?: string;
  hint?: string;
  character_count?: string;
  [k: string]: unknown;
}
export interface Icon {
  kind?: "icon";
  label?: string;
  src: string;
  decorative?: boolean;
  ariaLabel?: string;
  hint?: string;
}
export interface Custom {
  kind: "custom_background";
  blade_background_color?: string;
  blade_custom_image_desktop: Image;
  blade_custom_image_mobile: Image;
  blade_background_gradient_direction?: string;
}
export interface Standard {
  kind: "standard_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
export interface Swoosh {
  kind: "swoosh_background";
  blade_background_color: string;
}
export interface Container {
  kind: "container_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_image_position?: string;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
// ...
export interface CardResource {
  id: string;
  kind: "card_resource";
  post_id?: string | number;
  ucs_content_id?: ExternalRef;
  content_source_type?: "select-content" | "manual-content";
  content_library_type?: "wordpress" | "ucs";
  headline?: string;
  description?: string;
  badge?: string;
  link_1?: Cta;
  image?: Image;
  card_layout?: ("left" | "right" | "center")[];
  url?: string;
  taxonomy?: Taxonomy;
}
export interface ExternalRef {
  /**
   * content source
   */
  source?: string;
  /**
   * content identifier
   */
  id?: string;
  /**
   * additonal options for the control type and source type
   */
  typeOptions?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface Taxonomy {
  tagsAuto?: Tag[];
  tagsManual?: Tag[];
}
export interface Tag {
  uuid: string;
  tagName: string;
  tagUuid: string;
}
export interface CardHeadshot {
  id: string;
  kind: "card_headshot";
  customer_name: string;
  customer_attribution_role: string;
  customer_attribution_company: string;
  link_1?: Cta;
  image: Image;
}
export interface CardStory {
  id: string;
  kind: "card_story";
  headline?: string;
  description?: string;
  link_1?: Cta;
  image?: Image;
  taxonomy?: Taxonomy;
}
export interface CardQuote {
  id: string;
  kind: "card_quote";
  align_image?: "left" | "right";
  quote_image?: {
    headshot?: Image;
    logo?: Image;
    product?: Image;
    [k: string]: unknown;
  };
  quote_text: string;
  name: string;
  role: string;
  company_name: string;
  link_1?: Cta;
  card_background_type?: Background;
}
export interface CardStatistic {
  id: string;
  kind: "card_statistic";
  counter_animation?: boolean;
  statistic: string;
  caption: string;
  description?: string;
  link_1?: Cta;
  card_background_type?: TopBottomSplit;
}
export interface TopBottomSplit {
  kind: "top_bottom_split_background";
  background_image_top?: Image;
  background_image_bottom?: Image;
}
export interface CardEvent {
  id: string;
  kind: "card_event";
  image: Image;
  headline: string;
  date: string;
  location: string;
  description?: string;
  link_1?: Cta;
}
// ...
export interface CardWebinar {
  id: string;
  kind: "card_webinar";
  ucs_content_id?: ExternalRef;
  content_source_type?: "select-content";
  content_library_type?: "ucs";
  headline?: string;
  startDate?: string;
  endDate?: string;
  duration?: string;
  description?: string;
  badge?: string;
  link_1?: Cta;
  image?: Image;
  taxonomy?: Taxonomy;
}
// ...
export interface Nup {
  attributes?: {
    link_location: string;
    audience_segment?: string[];
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    card_type: string;
    card_layout?: string;
    grid_type?: string;
    card_description_visibility?: ToggleControl;
    anchor_id?: string;
    card_image?: ToggleControl;
    card_eyebrow_visibility?: ToggleControl;
    card_link_visibility?: ToggleControl;
    eyebrow?: Eyebrow;
    headline?: string;
    headline_font_color?: string;
    description?: string;
    link_1?: Cta;
    link_2?: Cta;
    legal?: string;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/nup";
  id: string;
  media_items?: (CardResource | CardHeadshot | CardQuote | CardStatistic | CardEvent | CardWebinar)[];
  [k: string]: unknown;
}